commit to pull

+++ GLOBALVARIABLES.h
@@ -4,5 +4,5 @@ float _WINDOW_H;
 float _WINDOW_W;
 
 float _DISPLAY_SCALE;
-float _KEYSIZE;
+double _KEYSIZE;
 float _TEXT_SCALE;
